Move paren_contents under ast and restrict syntax BUILD visibility
