+++ executable_semantics/ast/BUILD
@@ -28,11 +28,11 @@ cc_library(
     srcs = ["expression.cpp"],
     hdrs = ["expression.h"],
     deps = [
+        ":paren_contents",
         "//common:indirect_value",
         "//common:ostream",
         "//executable_semantics/common:arena",
         "//executable_semantics/common:error",
-        "//executable_semantics/syntax:paren_contents",
         "@llvm-project//llvm:Support",
     ],
 )
@@ -42,7 +42,7 @@ cc_test(
     srcs = ["expression_test.cpp"],
     deps = [
         ":expression",
-        "//executable_semantics/syntax:paren_contents",
+        ":paren_contents",
         "@llvm-project//llvm:gtest",
         "@llvm-project//llvm:gtest_main",
     ],
@@ -70,6 +70,11 @@ cc_library(
 )
 
 cc_library(
+    name = "paren_contents",
+    hdrs = ["paren_contents.h"],
+)
+
+cc_library(
     name = "pattern",
     srcs = ["pattern.cpp"],
     hdrs = ["pattern.h"],
@@ -86,8 +91,8 @@ cc_test(
     name = "pattern_test",
     srcs = ["pattern_test.cpp"],
     deps = [
+        ":paren_contents",
         ":pattern",
-        "//executable_semantics/syntax:paren_contents",
         "@llvm-project//llvm:Support",
         "@llvm-project//llvm:gtest",
         "@llvm-project//llvm:gtest_main",
