+++ executable_semantics/ast/expression_test.cpp
@@ -6,8 +6,8 @@
 
 #include <string>
 
+#include "executable_semantics/ast/paren_contents.h"
 #include "executable_semantics/common/arena.h"
-#include "executable_semantics/syntax/paren_contents.h"
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 #include "llvm/Support/Casting.h"
