+++ executable_semantics/ast/paren_contents.h
@@ -2,8 +2,8 @@
 // Exceptions. See /LICENSE for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#ifndef EXECUTABLE_SEMANTICS_SYNTAX_PAREN_CONTENTS_H_
-#define EXECUTABLE_SEMANTICS_SYNTAX_PAREN_CONTENTS_H_
+#ifndef EXECUTABLE_SEMANTICS_AST_PAREN_CONTENTS_H_
+#define EXECUTABLE_SEMANTICS_AST_PAREN_CONTENTS_H_
 
 #include <optional>
 #include <string>
@@ -83,4 +83,4 @@ auto ParenContents<Term>::TupleElements(int line_num) const
 
 }  // namespace Carbon
 
-#endif  // EXECUTABLE_SEMANTICS_SYNTAX_PAREN_CONTENTS_H_
+#endif  // EXECUTABLE_SEMANTICS_AST_PAREN_CONTENTS_H_
