+++ executable_semantics/syntax/parser.ypp
@@ -69,7 +69,7 @@
 #include "executable_semantics/ast/function_definition.h"
 #include "executable_semantics/ast/pattern.h"
 #include "executable_semantics/common/arena.h"
-#include "executable_semantics/syntax/paren_contents.h"
+#include "executable_semantics/ast/paren_contents.h"
 
 namespace Carbon {
 class ParseAndLexContext;
